@@ -1,6 +1,6 @@
 module github.com/tliron/candice
 
-go 1.15
+go 1.16
 
 // replace github.com/tliron/kutil => /Depot/Projects/RedHat/kutil
 
@@ -10,7 +10,7 @@ require (
 	github.com/openconfig/goyang v0.2.4
 	github.com/spf13/cobra v1.1.3
 	github.com/tebeka/atexit v0.3.0
-	github.com/tliron/kutil v0.1.20
+	github.com/tliron/kutil v0.1.21
 	github.com/tliron/yamlkeys v1.3.5
 	golang.org/x/crypto v0.0.0-20210220033148-5ea612d1eb83
 	gopkg.in/DATA-DOG/go-sqlmock.v1 v1.3.0 // indirect
